Properly handling multiple messages in the buffer.
I've changed the way I find lengths in messages and parse them out.
Before, if multiple messages ended up in the read buffer, we wouldn't
have noticed because we only try to parse when the amount of data read
exactly equals the length we were expecting.  Now, we can receive many
messages in the read buffer, and parse them out properly.