Making `Actor` implement `Runnable`.
Instead of extending `Thread`, we now implement `Runnable`.  Also, the
logic for each tick has been factored out into a `tick()` method.  This
means we can make `Actors` which *don't* run in their own thread.  Or,
by calling the `start()` method, we can start a new thread with this
Actor.