Properly handling key cancelling during selection.
The methods in the `Communicator` interface now return a boolean value.
This value is whether or not the key has been cancelled.  If it has, we
do *not* want to check to see if it's readable or writable in the
selection loop because this will throw a `CancelledKeyException` which
will halt the whole program.